pyproject.toml
[project]
name = "agentic-quiz-platform"
version = "0.1.0"
description = "Professional Agentic Quiz Platform using Google ADK"
readme = "README.md"
authors = [
    { name = "rehan363", email = "rehan007313@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.128.0",
    "google-adk>=1.18.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "litellm>=1.57.11",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[project.scripts]
quiz-platform = "app.main:app"

[tool.uv]
package = false
dev-dependencies = []

[tool.uv.sources]
app = { path = "src/app" }

